pyproject.toml
[project]
name = "py-ecommerce-synth-oltp"
version = "0.1.0"
description = "Generate synthetic data for an E-commerce OLTP system"
authors = [
    {name = "thuphan",email = "thuleanhphan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]

[tool.poetry]
packages = [{include = "mock_ecommerce", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
